chore: Move shared files out of prolly
