+++ src/array-compare.ts
