+++ src/peek-iterator.test.ts
