+++ src/peek-iterator.ts
